+++ src/apps/work/src/pages/challenges/ChallengeEditorPage/components/ReviewersField/ReviewConfigurationSummary.tsx
@@ -1,6 +1,5 @@
 import {
     FC,
-    useCallback,
     useEffect,
     useMemo,
     useState,
@@ -8,7 +7,6 @@ import {
 import classNames from 'classnames'
 
 import {
-    useFetchChallengeReviewContext,
     useFetchResourceRoles,
     useFetchResources,
 } from '../../../../../lib/hooks'
@@ -41,6 +39,7 @@ import {
 import {
     buildAssignedResourcesByReviewer,
 } from './reviewerAssignments.utils'
+import { ReviewContextSection } from './ReviewContextSection'
 import {
     isAiReviewer,
     normalizeReviewerText,
@@ -515,51 +514,6 @@ export const ReviewConfigurationSummary: FC<ReviewConfigurationSummaryProps> = (
             workflowMap,
         ],
     )
-    const reviewContextResult = useFetchChallengeReviewContext(props.challengeId)
-    const reviewContextRequirements = useMemo(() => {
-        const requirements = reviewContextResult.context?.context?.requirements
-
-        if (!Array.isArray(requirements)) {
-            return []
-        }
-
-        return [...requirements].sort((a, b) => a.id.localeCompare(b.id))
-    }, [reviewContextResult.context])
-    const reviewContextLoading = reviewContextResult.isLoading
-    const reviewContextError = reviewContextResult.error
-    const hasReviewContextRequirements = reviewContextRequirements.length > 0
-    const reviewContextSectionIsVisible = Boolean(props.challengeId)
-    const [expandedReviewContextIds, setExpandedReviewContextIds] = useState<string[]>([])
-
-    useEffect(() => {
-        if (reviewContextLoading || reviewContextError) {
-            return
-        }
-
-        setExpandedReviewContextIds(
-            reviewContextRequirements.length > 0
-                ? [reviewContextRequirements[0].id]
-                : [],
-        )
-    }, [reviewContextError, reviewContextLoading, reviewContextRequirements])
-
-    const handleToggleRequirement = useCallback((requirementId: string): void => {
-        setExpandedReviewContextIds(prev => {
-            if (prev.includes(requirementId)) {
-                return prev.filter(id => id !== requirementId)
-            }
-
-            return [...prev, requirementId]
-        })
-    }, [])
-
-    const handleRetryReviewContext = useCallback(async (): Promise<void> => {
-        if (!reviewContextResult.mutate) {
-            return
-        }
-
-        await reviewContextResult.mutate()
-    }, [reviewContextResult])
 
     const referencedScorecardIds = useMemo(
         () => getReferencedScorecardIds(humanReviewers, workflowsToDisplay, workflowMap),
@@ -1119,131 +1073,7 @@ export const ReviewConfigurationSummary: FC<ReviewConfigurationSummaryProps> = (
                 )
                 : undefined}
 
-            {reviewContextSectionIsVisible
-                ? (
-                    <section className={classNames(styles.card, styles.reviewContextSection)}>
-                        <div className={styles.cardHeader}>
-                            <span aria-hidden='true' className={styles.headerIcon}>📋</span>
-                            <h5 className={styles.cardTitle}>
-                                Review Context Requirements
-                                {' '}
-                                (
-                                {reviewContextRequirements.length}
-                                )
-                            </h5>
-                        </div>
-                        <div className={styles.cardBody}>
-                            {reviewContextLoading
-                                ? (
-                                    <div className={styles.emptyState}>
-                                        Loading review context requirements...
-                                    </div>
-                                )
-                                : reviewContextError
-                                    ? (
-                                        <div className={styles.reviewContextErrorState}>
-                                            <div className={styles.errorMessage}>
-                                                {reviewContextError}
-                                            </div>
-                                            <button
-                                                type='button'
-                                                className={styles.retryButton}
-                                                onClick={handleRetryReviewContext}
-                                            >
-                                                Retry
-                                            </button>
-                                        </div>
-                                    )
-                                    : hasReviewContextRequirements
-                                        ? (
-                                            <div className={styles.reviewContextList}>
-                                                {reviewContextRequirements.map(requirement => {
-                                                    const priority = normalizeReviewerText(requirement.priority) || 'medium'
-                                                    const priorityClass = priority === 'high'
-                                                        ? styles.priorityHigh
-                                                        : priority === 'low'
-                                                            ? styles.priorityLow
-                                                            : styles.priorityMedium
-                                                    const isExpanded = expandedReviewContextIds.includes(requirement.id)
-
-                                                    return (
-                                                        <div
-                                                            className={styles.reviewContextCard}
-                                                            key={requirement.id}
-                                                        >
-                                                            <button
-                                                                type='button'
-                                                                className={styles.reviewContextCardHeader}
-                                                                onClick={function (): void {
-                                                                    handleToggleRequirement(requirement.id)
-                                                                }}
-                                                                aria-expanded={isExpanded}
-                                                            >
-                                                                <span className={styles.reviewContextCardToggle}>
-                                                                    {isExpanded ? '▼' : '▶'}
-                                                                </span>
-                                                                <span className={styles.requirementId}>
-                                                                    [
-                                                                    {requirement.id}
-                                                                    ]
-                                                                </span>
-                                                                <span className={classNames(
-                                                                    styles.priorityBadge,
-                                                                    priorityClass,
-                                                                )}
-                                                                >
-                                                                    {priority.toUpperCase()}
-                                                                </span>
-                                                                <span className={styles.reviewContextTitle}>
-                                                                    {requirement.title}
-                                                                </span>
-                                                            </button>
-                                                            <div
-                                                                className={styles.reviewContextCardBody}
-                                                                hidden={!isExpanded}
-                                                            >
-                                                                <p className={styles.reviewContextDescription}>
-                                                                    {requirement.description}
-                                                                </p>
-                                                                {Array.isArray(requirement.constraints)
-                                                                    && requirement.constraints.length > 0
-                                                                    ? (
-                                                                        <ul className={styles.reviewContextConstraints}>
-                                                                            {[...requirement.constraints]
-                                                                                .sort((a, b) => a.id.localeCompare(b.id))
-                                                                                .map(constraint => (
-                                                                                    <li
-                                                                                        className={styles.reviewContextConstraintItem}
-                                                                                        key={constraint.id}
-                                                                                    >
-                                                                                        <span className={styles.constraintId}>
-                                                                                            [
-                                                                                            {constraint.id}
-                                                                                            ]
-                                                                                        </span>
-                                                                                        <span className={styles.constraintText}>
-                                                                                            {constraint.text}
-                                                                                        </span>
-                                                                                    </li>
-                                                                                ))}
-                                                                        </ul>
-                                                                    )
-                                                                    : undefined}
-                                                            </div>
-                                                        </div>
-                                                    )
-                                                })}
-                                            </div>
-                                        )
-                                        : (
-                                            <div className={styles.emptyState}>
-                                                No review context requirements defined.
-                                            </div>
-                                        )}
-                        </div>
-                    </section>
-                )
-                : undefined}
+            <ReviewContextSection challengeId={props.challengeId} />
 
             {humanReviewers.length
                 ? (
